Tirando o fundo branco da setinha.

+++ src/base/audioRepository.cxx
@@ -73,7 +73,7 @@ AudioRepository::AudioRepository() {
  }
 
  //quando quiser tocar um som é só chamar este método passando o nome
-  void AudioRepository::play_sound(const string& name, bool unique, int volume){
+  void AudioRepository::play_sound(const string& name, bool unique, float volume){
 	 if (unique) {
 		 if (playing) return;
